Update overview.md
QuickStart is creating world-class technologists by personalizing and individualizing training to address the massive skills gap in the IT industry. Through 20 years of research and data analysis, we’ve learned that a modern learner prefers to learn through multiple modalities such as self-paced online learning, virtual instructor-led training, on-the-job training, implementation, and hands-on experience through labs, mentoring, and social and peer-to-peer learning. We packaged all the modalities into a single cloud-based platform called CLIPP.

Recently, we have introduced this new course on "Kubernetes Fundamentals". This course will take our students from zero to knowing how to deploy a containerized application and manipulate resources via the API. This course is designed to work with a wide range of Linux distributions, enabling our students to apply these concepts regardless of their distro.